requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.56.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
pandas==2.1.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
